projects/Django-Scrapy-Selenium/requirements/base.txt
python-slugify==8.0.4  # https://github.com/un33k/python-slugify
Pillow==11.1.0  # https://github.com/python-pillow/Pillow
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.8.2  # https://github.com/evansd/whitenoise
redis==5.2.1  # https://github.com/redis/redis-py
hiredis==3.1.0  # https://github.com/redis/hiredis-py
celery==5.4.0  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.7.0  # https://github.com/celery/django-celery-beat
flower==2.0.1  # https://github.com/mher/flower

# Django
# ------------------------------------------------------------------------------
django==5.1.5  # pyup: < 5.1  # https://www.djangoproject.com/
# django==5.0.10  # pyup: < 5.1  # https://www.djangoproject.com/
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-model-utils==5.0.0  # https://github.com/jazzband/django-model-utils
django-allauth[mfa]==65.3.1  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.3  # https://github.com/django-crispy-forms/django-crispy-forms
# crispy-bootstrap5==2024.10  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.4.0  # https://github.com/jazzband/django-redis
django-webpack-loader==3.1.1  # https://github.com/django-webpack/django-webpack-loader
django-celery-results
crispy-tailwind
django-filter
django-simple-captcha
django-htmx
django-widget-tweaks
django-tables2
django-render-block
django-admin-sortable2
django-formtools
django-formset-js-improved
django-import-export
django-unfold

scrapy-redis
celery-progress
dateparser
olefile
django-modeltranslation
django-dynamic-formsets
bs4
django-ckeditor
jwt
django-template-partials

Scrapy
scrapy-selenium==0.0.7
# selenium==3.9.0
selenium==4.28.1
typing-extensions
webdriver-manager
